% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BinomialProbability.R
\name{binomprob}
\alias{binomprob}
\title{Binomial Proability.}
\usage{
binomprob(k, n, prob, lower.tail)
}
\arguments{
\item{k}{is the number of successes of interest (must be integer).}

\item{n}{the number of trials.}

\item{prob}{probabilty of success}

\item{lower.tail}{is a Boolean for finding the probability above (FALSE) or below (TRUE) the inputted value (inclusive)}
}
\value{
Right or left tail Probabilty of random variable X
}
\description{
Binomial Proability.
}
\examples{
binomprob(3,20,1/3,lower.tail = FALSE)
binomprob(3,20,1/3,lower.tail = TRUE)
}

