% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExactBinomTestAndCI.R
\name{binomtest}
\alias{binomtest}
\title{Exact Binomial Test and Confidence Interval.}
\usage{
binomtest(observed, n, hypothesized = NULL, alternative, conf.level = NULL)
}
\arguments{
\item{observed}{number of success.}

\item{n}{sample size.}

\item{hypothesized}{probability of success.}

\item{alternative}{probability of success.}

\item{conf.level}{confidence level for a two-sided confidence interval.}
}
\value{
Exact Binomial Test
}
\description{
Exact Binomial Test and Confidence Interval.
}
\examples{
binomtest(3, 13, 1/2, "greater")
binomtest(3, 13, 1/2, "less")
binomtest(3, 13, 1/2, "two.sided")
binomtest(3, 13, conf.level=c(90,95))
}

