% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImproveISCAM.R
\docType{package}
\name{ImproveISCAM-package}
\alias{ImproveISCAM-package}
\title{The most incredible editable ImproveISCAM package}
\description{
Improve functions for ISCAM Book.
}
\details{
Probability Desired.
}
\section{Indy}{
 This is not working but you will make it work.
}
\keyword{package}

