% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalProbability.R
\name{normprob}
\alias{normprob}
\title{Normal Probability}
\usage{
normprob(xval, mean = 0, sd = 1, direction, label = NULL, xval2 = NULL)
}
\arguments{
\item{direction}{is a string for finding the probability above,below, between, or outside}

\item{label}{the horizontal axis}

\item{xval2}{}
}
\value{
Normal Probability
}
\description{
Normal Probability
}
\examples{
normprob(2, direction="below")
normprob(2, direction="above")
normprob(-2, direction="between",xval2=2)
normprob(-2, direction="outside",xval2=2)
}

