% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZtestCI.R
\name{propztest}
\alias{propztest}
\title{z-test for proportion and confidence interval}
\usage{
propztest(observed, n, hypothesized = NULL, alternative = "two.sided",
  conf.level = NULL)
}
\arguments{
\item{observed}{number of successes or sample proportion}

\item{n}{sample size}

\item{hypothesized}{probability of successes}

\item{alternative}{less, greater, or two.sided}

\item{conf.level}{for a two-sided confidence interval}
}
\value{
z test result
}
\description{
z-test for proportion and confidence interval
}
\examples{
propztest(3, 13, 1/2, "greater")
propztest(3, 13, 1/2, "less")
propztest(3, 13, 1/2, "two.sided")
propztest(3, 13, conf.level=c(90,95))
}

